+++ mantra/transforms/util_transforms.py
@@ -1,7 +1,6 @@
 from itertools import combinations
 from typing import Dict, List, Literal, TypeAlias, Union
 
-import numpy as np
 import torch
 from torch import Tensor
 from torch_geometric.data import Data
@@ -121,6 +120,8 @@ class SelectFeatures(BaseTransform):
         assert isinstance(
             src_tensor, Dict
         ), f"The attribute {self.src} is not of type dict"
+        assert self.dst is not None, "`dst` is None"
+
         assert len(self.dst) == len(
             src_tensor.keys()
         ), f"There is a mismatch between num of `src` keys ( {len(src_tensor.keys())} ) and `dst` targets ( { len(self.dst) } )"
@@ -209,6 +210,10 @@ class PropagateConvexComb(BaseTransform):
             self.source in data
         ), f"Data object is missing source tensor `{self.source}`"
 
+        assert isinstance(
+            getattr(data, self.source), torch.Tensor
+        ), f"Input {self.source} is not a torch.Tensor"
+
         x = getattr(data, self.source)
         triangulation = getattr(data, "triangulation")
 
@@ -230,21 +235,15 @@ class PropagateConvexComb(BaseTransform):
 
         for dim in range(1, max_dim):
             simplices_ = [s for s in simplices if len(s) == dim + 1]
-            M = []
 
-            for s in simplices_:
-                # View as an array to correct for the index shift; our
-                # triangulation is not zero-indexed.
-                s = np.asarray(s)
+            idx = torch.tensor(simplices_) - 1
 
-                # Calculate barycenter for the current simplex (i.e., one
-                # row of the result matrix).
-                # TODO: Change this to another type of combination function
-                M.append(np.mean(x[s - 1, :], axis=0))
+            # Calculate all barycenters of the current rank at once.
+            # TODO: Change this to another type of combination function
+            values[f"x_{dim}"] = x[idx].mean(dim=1)
 
-            M = np.asarray(M)
-            values[f"x_{dim}"] = torch.from_numpy(M).to(torch.float32)
-
+        # Assignment to data object
         for k, v in values.items():
             data[k] = v
+
         return data